Create components folder
