+++ src/components/Container.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const Container = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: flex-end;
+`
