+++ src/components/Item.tsx
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const Item = styled.div`
+  border: 2px dashed red;
+`
