+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './Input'
+export * from './Container'
+export * from './Item'
